se actualizaron campos int a big int

+++ database/migrations/2024_09_19_000002_create_products_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->string('name', 100);
             $table->text('description')->nullable();
             $table->decimal('price', 10, 2);
-            $table->integer('stock')->default(0);
+            $table->bigInteger('stock')->default(0);
             $table->timestamp('created_at')->useCurrent();
         });
     }
