+++ database/migrations/2024_09_19_000004_create_invoice_items_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('invoice_id')->constrained()->onDelete('cascade');
             $table->foreignId('product_id')->constrained()->onDelete('cascade');
-            $table->integer('quantity');
+            $table->bigInteger('quantity');
             $table->decimal('price', 10, 2);
         });
     }
